requirements.txt
# Django & DRF
Django>=5.1,<6.0
djangorestframework>=3.15.0
django-cors-headers>=4.3.0

# Database
psycopg[binary]>=3.1.12

# Celery & Redis (for scheduled crawling)
celery>=5.4.0
redis>=5.0.0
django-celery-beat>=2.6.0
django-celery-results>=2.5.0

# ML & NLP
scikit-learn>=1.4.0
nltk>=3.8.1
numpy>=1.26.0
pandas>=2.2.0

# Crawler
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
selenium>=4.15.0
webdriver-manager>=4.0.0
undetected-chromedriver>=3.5.0
setuptools>=69.0.0

# Utilities
python-dotenv>=1.0.0

# Testing
pytest>=8.0.0
pytest-django>=4.8.0
